feat(resx): use rusted_realm extractor
This commit switches from using JPEXS to extract the packets to using
the rusted_realm extractor lib. This change allows the resx API to be a
lot simpler.

BREAKING CHANGE: The methods `makeGSCPath`, `unpackSwf` and
`extractPacketInfo` have been replaced by a single method
`extractPackets`.

This new method takes a path to a RotMG client swf as the argument and
returns a promise which resolves with a PacketMap object that contains
the extracted packet ids.