chore(resx): improve extractPackets
The extractPackets function now makes use of the new WASM binary for
extracting packets. This simplifies the method a lot, and also avoids
the need to save the client to the disk since the method now takes a
Buffer containing the client as its argument.

BREAKING CHANGE: code which uses the extractPackets method will need to
be updated to pass the client buffer instead of a path to the client.